watchers: rework thread handling by synchronizing with a queue of events
Instead of synchronizing the various threads with a condition and
executing code in them, we use a queue to send events to a FSM in the
main thread. Zookeeper and Observer thread callbacks will just post an
event in the event queue.

The code running in non-main threads is therefore minimal (only
posting events). Locks are not needed anymore. Handling connection
lost is also far more robust. Error handling is also simplified by
reposting failed events in case of Zookeeper errors.