joiner: add an option to work with a common node
In normal operations, each joiner will have its own zookeeper node. We
add an option to make them work on a common node instead. The hostname
is not added to the node and local modifications are not
authoritative. Moreover, the node is not ephemeral. Instead of using
an IP address, `common` is used as the node name.