zkfarmer: add `zkfarmer import` command
This commands will import local configuration to ZooKeeper but will
not update the local configuration with ZooKeeper content. This is not
the exact reverse of `zkfarmer export` since it is only able to import
one node. This is a limited version of `zkfarmer join` where changes
from ZooKeeper are ignored.

This is useful when you have a large farm of joinder using a common
node to sync configuration.